# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Raudah Travels & Tours API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agents/public-apply:
    post:
      operationId: publicApplyAsAgent
      tags: [agents]
      summary: Submit public agent application (no auth required)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublicAgentApplicationBody"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicApplyResponse"

  # --- Agent wallet ---
...
  /agents/profile:
    get:
      operationId: getAgentProfile
      tags: [agents]
      summary: Get agent profile for current user
      responses:
        "200":
          description: Agent profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentProfile"
    put:
      operationId: updateAgentProfile
      tags: [agents]
      summary: Update agent profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAgentProfileBody"
      responses:
        "200":
          description: Updated agent profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentProfile"
  /agents:
    get:
      operationId: listAgents
      tags: [agents]
      summary: List all agents (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
        - name: limit
          in: query
          schema:
            type: number
        - name: offset
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of agents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentListResponse"
  /agents/apply:
    post:
      operationId: applyAsAgent
      tags: [agents]
      summary: Submit agent application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentApplicationBody"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentProfile"
  /agents/{id}/approve:
    put:
      operationId: approveAgent
      tags: [agents]
      summary: Approve or reject agent application (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveAgentBody"
      responses:
        "200":
          description: Agent updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentProfile"
  /agents/commissions:
    get:
      operationId: listCommissions
      tags: [agents]
      summary: List commissions for current agent
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, paid]
        - name: limit
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of commissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommissionListResponse"

  # --- Dashboard (wow endpoints) ---
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        clerkUserId:
          type: string
        email:
          type: string
        fullName:
          type: string
        phone:
          type: string
        avatarUrl:
          type: string
          nullable: true
        role:
          type: string
          enum: [user, agent, staff, admin, super_admin]
        ninNumber:
          type: string
          nullable: true
        passportNumber:
          type: string
          nullable: true
        dateOfBirth:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, clerkUserId, email, fullName, role, createdAt]
...
    TravelPackage:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [hajj, umrah]
        category:
          type: string
          enum: [premium, standard, budget]
        season:
          type: string
          nullable: true
        description:
          type: string
        price:
          type: number
        agentDiscount:
          type: number
        depositAmount:
          type: number
        durationDays:
          type: number
        departureDate:
          type: string
        returnDate:
          type: string
        departureCities:
          type: array
          items:
            type: string
        airlines:
          type: array
          items:
            type: string
        maxCapacity:
          type: number
        currentBookings:
          type: number
        inclusions:
          type: array
          items:
            type: string
        imageUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        starRating:
          type: number
        status:
          type: string
          enum: [active, draft, archived]
        countdownEnabled:
          type: boolean
        countdownExpiry:
          type: string
          nullable: true
        countdownAction:
          type: string
          enum: [disable, both]
        isRegistrationClosed:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - type
        - category
        - description
        - price
        - agentDiscount
        - depositAmount
        - durationDays
        - departureDate
        - returnDate
        - departureCities
        - airlines
        - maxCapacity
        - currentBookings
        - inclusions
        - isActive
        - isFeatured
        - starRating
        - status
        - countdownEnabled
        - countdownAction
        - isRegistrationClosed
        - createdAt
...
    Booking:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        packageId:
          type: string
        agentId:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
        totalPrice:
          type: number
        amountPaid:
          type: number
        pilgrimCount:
          type: number
        pilgrimDetails:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
        package:
          $ref: "#/components/schemas/TravelPackage"
          nullable: true
        user:
          $ref: "#/components/schemas/UserProfile"
          nullable: true
      required: [id, userId, packageId, status, totalPrice, amountPaid, pilgrimCount, createdAt, updatedAt]
...
    AgentProfile:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        businessName:
          type: string
        contactPerson:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        agentCode:
          type: string
          nullable: true
        commissionRate:
          type: number
        commissionType:
          type: string
          enum: [percentage, fixed]
        walletBalance:
          type: number
        isVerified:
          type: boolean
        status:
          type: string
          enum: [pending, active, suspended]
        bio:
          type: string
          nullable: true
        createdAt:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
          nullable: true
      required: [id, userId, businessName, commissionRate, commissionType, walletBalance, isVerified, status, createdAt]
    UpdateAgentProfileBody:
      type: object
      properties:
        businessName:
          type: string
        contactPerson:
          type: string
        email:
          type: string
        phone:
          type: string
        bio:
          type: string
    AgentApplicationBody:
      type: object
      properties:
        businessName:
          type: string
        contactPerson:
          type: string
        email:
          type: string
        phone:
          type: string
        bio:
          type: string
      required: [businessName]
    ApproveAgentBody:
      type: object
      properties:
        status:
          type: string
          enum: [approved, rejected]
        commissionRate:
          type: number
      required: [status]
    AgentListResponse:
      type: object
      properties:
        agents:
          type: array
          items:
            $ref: "#/components/schemas/AgentProfile"
        total:
          type: number
      required: [agents, total]
    Commission:
      type: object
      properties:
        id:
          type: string
        agentId:
          type: string
        bookingId:
          type: string
        amount:
          type: number
        status:
          type: string
          enum: [pending, paid]
        createdAt:
          type: string
        booking:
          $ref: "#/components/schemas/Booking"
          nullable: true
      required: [id, agentId, bookingId, amount, status, createdAt]
    CommissionListResponse:
      type: object
      properties:
        commissions:
          type: array
          items:
            $ref: "#/components/schemas/Commission"
        totalEarned:
          type: number
        pendingAmount:
          type: number
      required: [commissions, totalEarned, pendingAmount]
...
    PublicAgentApplicationBody:
      type: object
      properties:
        businessName:
          type: string
        contactPerson:
          type: string
        email:
          type: string
        phone:
          type: string
        bio:
          type: string
          nullable: true
        experienceYears:
          type: number
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
      required: [businessName, contactPerson, email, phone]
    PublicApplyResponse:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
        message:
          type: string
      required: [id, status, message]
...